Refactored Character class to inherit from sf::Sprite

+++ include/Projectile.h
@@ -16,14 +16,14 @@ class Projectile {
 public:
     int damage;
     int unsigned speed, wait, type;
-    double x, y, ax, ay, orient;
+    float x, y, ax, ay, orient;
     sf::Sprite C;
 
     Character *orig;
 
     void setorig(Character *orig);
 
-    double rotatexy(int mx, int my);
+    double rotatexy(float mx, float my);
 
     bool canWalk(FloorTile floortile[n][m], Station &game);
     //Projectile();
